requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.5
python-dotenv==1.0.0
PyPDF2==3.0.1
gunicorn==21.2.0